+++ runtimes/bulletin-paseo/src/lib.rs
@@ -30,8 +30,6 @@ pub mod fast_runtime_binary {
 mod genesis_config_presets;
 pub mod paseo_constants;
 pub mod storage;
-#[cfg(test)]
-mod tx_pause_tests;
 mod weights;
 pub mod xcm_config;
